+++ src/utils/openwhisk.js
@@ -16,9 +16,18 @@ const owwrapper = require('@adobe/openwhisk-loggly-wrapper');
 
 /**
  * Builds the request path from path, selector, extension and params
- * @param {Object} p The request parameters
+ * @param {Object} req The action request
  */
-function buildPath(p) {
+function buildPath(req) {
+  // workaround for https://github.com/adobe/helix-pipeline/issues/254 until `contentRoot` is
+  // supplied. check for `x-old-url` header.
+  if (req.headers && req.headers['x-old-url']) {
+    const u = req.headers['x-old-url'];
+    const idx = u.indexOf('?');
+    return idx > 0 ? u.substring(0, idx) : u;
+  }
+
+  const p = req.params;
   const rootPath = p.rootPath || '';
   let path = p.path || '/';
   const dot = path.lastIndexOf('.');
@@ -41,7 +50,7 @@ function extractClientRequest(action) {
   if (!request || !request.params) {
     return {};
   }
-  const reqPath = buildPath(request.params);
+  const reqPath = buildPath(request);
   const query = request.params.params ? `?${request.params.params}` : '';
   return {
     // the edge encodes the client request parameters into the `params` param ;-)
@@ -63,17 +72,17 @@ function extractClientRequest(action) {
 async function createActionResponse(payload) {
   const {
     response: {
-      status = 200,
+      status,
       headers = { 'Content-Type': 'application/json' },
       body = headers['Content-Type'] === 'application/json' ? {} : '',
-      error,
     } = {},
+    error,
   } = payload;
   return {
-    statusCode: status,
+    statusCode: status || (error ? 500 : 200),
     headers,
     body,
-    error,
+    error: error ? (error.stack || error) : undefined, // workaround for https://github.com/apache/incubator-openwhisk-runtime-nodejs/issues/63
   };
 }
 
@@ -101,6 +110,9 @@ function extractActionContext(params) {
       delete disclosed[key];
     }
   });
+
+  // extract content (will be added to payload)
+  delete disclosed.content;
 
   // setup action
   return {
@@ -131,6 +143,10 @@ async function runPipeline(cont, pipe, actionParams) {
     const payload = {
       request: extractClientRequest(action),
     };
+    if (params.content) {
+      // pass content param from request to payload
+      payload.content = params.content;
+    }
     return pipe(cont, payload, action);
   }
   return createActionResponse(await owwrapper(runner, actionParams));