+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## [1.3.3](https://github.com/adobe/helix-pipeline/compare/v1.3.2...v1.3.3) (2019-04-12)
+
+
+### Bug Fixes
+
+* **openwhisk:** context.request.path & context.request.url are incorrect for subdirectory strains ([#256](https://github.com/adobe/helix-pipeline/issues/256)) ([7fce752](https://github.com/adobe/helix-pipeline/commit/7fce752)), closes [#254](https://github.com/adobe/helix-pipeline/issues/254)
+
+## [1.3.2](https://github.com/adobe/helix-pipeline/compare/v1.3.1...v1.3.2) (2019-04-12)
+
+
+### Bug Fixes
+
+* **openwhisk:** properly propagate error in action response ([#238](https://github.com/adobe/helix-pipeline/issues/238)) ([c680eed](https://github.com/adobe/helix-pipeline/commit/c680eed)), closes [#237](https://github.com/adobe/helix-pipeline/issues/237)
+
+## [1.3.1](https://github.com/adobe/helix-pipeline/compare/v1.3.0...v1.3.1) (2019-04-11)
+
+
+### Bug Fixes
+
+* **package:** update @adobe/helix-shared to version 0.10.3 ([ce30189](https://github.com/adobe/helix-pipeline/commit/ce30189))
+
+### Features
+
+* **openwhisk**: add request body to payload #235
+
+
 # [1.3.0](https://github.com/adobe/helix-pipeline/compare/v1.2.4...v1.3.0) (2019-04-10)